Event creation now fully working with styling and backend implementation
