+++ client/src/assets/jss/components/eventStyle.js
@@ -1,0 +1,19 @@
+import './../../fonts/fonts.css';
+
+const eventContainer = {
+    width: '50%',
+    margin: '0 auto',
+    top: '50%',
+    transform: 'translateY(50%)',
+    fontFamily: 'Poppins'
+}
+
+const eventForm = {
+    marginTop: '5%'
+}
+
+const calendarSpacing = {
+    marginRight: '2%'
+}
+
+export { eventContainer, eventForm, calendarSpacing };
