+++ client/src/components/Login/Login.js
@@ -4,18 +4,8 @@ import { Form, Button } from "react-bootstrap";
 import { Redirect } from "react-router-dom";
 import { AUTHENTICATED, ADMIN, TOKEN } from "../../constants/sessionstorage";
 import { API_URL } from "../../constants/apiurl";
-import {
-  limiter,
-  loginContainer,
-  loginWrapper,
-  loginTitle,
-  inputWrapper,
-  buttonWrapper,
-  button100,
-  button100Wrapper,
-  inputStyle,
-  invalidInput
-} from "../../assets/jss/components/loginStyle";
+import { title, button100, buttonWrapper, button100Wrapper, invalidInput } from "../../assets/jss/sharedStyling";
+import { limiter, loginContainer, loginWrapper, inputWrapper, inputStyle } from "../../assets/jss/components/loginStyle";
 
 class Login extends React.Component {
   constructor(props) {
@@ -48,7 +38,7 @@ class Login extends React.Component {
         <div style={loginContainer}>
           <div style={loginWrapper}>
             <Form>
-              <span style={loginTitle}> Iniciar sesión </span>
+              <span style={title}> Iniciar sesión </span>
 
               {this.state.isInvalid ? this.renderWrongDataMessage() : null}
